% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphe.R
\docType{methods}
\name{allShortestPaths,Graph-method}
\alias{allShortestPaths,Graph-method}
\title{fonction qui crée le grpah pour les calculs de distances, ne peut
pas etre une methode de la classe Graphe cause des erreurs de profondeurs
de calculs}
\usage{
\S4method{allShortestPaths}{Graph}(object, data)
}
\arguments{
\item{data}{frame des reactions extraites du KGML en listes}
}
\description{
fonction qui crée le grpah pour les calculs de distances, ne peut
pas etre une methode de la classe Graphe cause des erreurs de profondeurs
de calculs
}
\examples{
aLLShortestPaths(igraph, data)
}
\keyword{kegg}

