% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphCreation.R
\name{createGraph}
\alias{createGraph}
\title{creation of graphe with iGraph
with the list of compounds (nodes) in the KGML
and the list of reations related to a gene}
\usage{
createGraph(pathwayId)
}
\description{
creation of graphe with iGraph
with the list of compounds (nodes) in the KGML
and the list of reations related to a gene
}

