% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphe.R
\docType{methods}
\name{createIGraph,GraphElements-method}
\alias{createIGraph,GraphElements-method}
\title{Creation of object igraph with data from XML file of pathway}
\usage{
\S4method{createIGraph}{GraphElements}(object)
}
\arguments{
\item{GraphElements}{}
}
\description{
Creation of object igraph with data from XML file of pathway
}
\examples{
createIGraph(GraphElements)
}
\keyword{graphElements}

