% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphe.R
\docType{methods}
\name{fromDFEntryToIGraphIdDF,Graph-method}
\alias{fromDFEntryToIGraphIdDF,Graph-method}
\title{A igraph function}
\usage{
\S4method{fromDFEntryToIGraphIdDF}{Graph}(object, data, indiceMetabolite)
}
\arguments{
\item{graphe,}{kegg id of compound of interest, dataframe of compounds
(nodes)}
}
\description{
Modification of dataframe giving in entry by user, from
c("gene", "metabolite") to 2 dataFrame
if indiceMetabolite = 1
d1 = c("metabolite1giIdIniGraph", "metabolitesIdIniGraph")
if indiceMetabolite = 2
d2 = c("metabolite2giIdIniGraph", "metabolitesIdIniGraph")
}
\examples{
fromDFEntryToIGraphIdDF(g, data, indiceMetabolite)
}
\keyword{igraph,}
\keyword{kgmlId}
\keyword{node,}

