% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphe.R
\name{getAllShortestDistances}
\alias{getAllShortestDistances}
\title{Fonction that calculates every shortest distances between each gene and
all metabolites.}
\usage{
getAllShortestDistances(pathwayId, data)
}
\arguments{
\item{data}{frame des reactions extraites du KGML en listes}
}
\description{
Fonction that calculates every shortest distances between each gene and
all metabolites.
}
\examples{
getAllShortestDistances(igraph, data)
}
\keyword{kegg}

