% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseXML.R
\name{getListEdgeFromOrthologKGML}
\alias{getListEdgeFromOrthologKGML}
\title{A KEGG Function}
\usage{
getListEdgeFromOrthologKGML(pathwayId)
}
\arguments{
\item{pathway}{id}
}
\description{
This function extract from KGML a list of reaction of entry-type "ortholog"
Returns a dataFrame object : id, entryId, reaction, ko
}
\examples{
getListEdgeFromOrthologKGML("hsa01100")
}
\keyword{Ortholog,}
\keyword{Reaction}
\keyword{kegg,}

