% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{getShortestDistanceBetween2Compounds}
\alias{getShortestDistanceBetween2Compounds}
\title{A graph function}
\usage{
getShortestDistanceBetween2Compounds(g, c1, c2)
}
\arguments{
\item{graphe,}{kegg id of compound of interest, dataframe of compounds
(nodes)}
}
\description{
fucntion return distance between 2 compound of the graph, in our case
the first (c1) compound is one of the edge of our gene of interest and the second
(c2) is the target metabolite
}
\examples{
getCompoundNodeKgmlId(g, compoundKeggId, nodeDF)
}
\keyword{graph,}
\keyword{kgmlId}
\keyword{node,}

