% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphe.R
\name{heatmapFunction}
\alias{heatmapFunction}
\title{Fonction to barplot all the shortest paths between 1 gene of interest to all
metabolites in the data value in entry}
\usage{
heatmapFunction(pathwayId, data)
}
\arguments{
\item{keggpathwayId,}{data.frame() colunm c(gene, metabolites) representing association
between them for each row, gene of interest}
}
\description{
Fonction to barplot all the shortest paths between 1 gene of interest to all
metabolites in the data value in entry
}
\examples{
heatmapFunction(hsa01100, data)
}
\keyword{barplot,}
\keyword{igraphe,}
\keyword{shortestDistance}

