% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermutationFunction.R
\name{permutationFunction}
\alias{permutationFunction}
\title{Permutation test to evauluate if gene-metabolite associations pairs are significantly closer
than randomly selected gene-metabolite pairs}
\usage{
permutationFunction(pathwayId, data, geneMeasured, metaboliteMeasured,
  permutation, output = c("medians", "pvalue", "histogram"))
}
\arguments{
\item{data, }{data.frame() colunm c(gene, metabolite) representing association
between them for each row geneMeasured, list of evaluated genes in study
(associated and not associated) metaboliteMeasured, list of measured
metabolites in study (associated and not associated)
permutation, number of permutation to execute
output, "medians" median value of permutated sets - "pvalue" - "histogram"}
}
\description{
Permutation test to evauluate if gene-metabolite associations pairs are significantly closer
than randomly selected gene-metabolite pairs
}
\examples{
permutationFunction("hsa01100", data, gene, metabolite, 1000)
}
\keyword{permutation}

