% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgeListFunctions.R
\name{unlistEdgeReactionCol}
\alias{unlistEdgeReactionCol}
\title{Fonction qui sépare les listes de réactions : id R1... R2... R3... ko
en plusieurs lignes avec les même id et le même ko.
-> id R1 ko
-> id R2 ko
-> id R3 ko}
\usage{
unlistEdgeReactionCol(edgeDataFrame)
}
\arguments{
\item{data}{frame des reactions extraites du KGML en listes}
}
\description{
Fonction qui sépare les listes de réactions : id R1... R2... R3... ko
en plusieurs lignes avec les même id et le même ko.
-> id R1 ko
-> id R2 ko
-> id R3 ko
}
\examples{
completeEdgeList(edgeDataFrame)
}
\keyword{kegg}

