% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphe.R
\name{barplotFunction}
\alias{barplotFunction}
\title{Fonction to barplot all the shortest paths between 1 gene of interest to all
metabolites in the data value in entry}
\usage{
barplotFunction(pathwayId, data, gene)
}
\arguments{
\item{keggpathwayId,}{data.frame() colunm c(gene, metabolites) representing association
between them for each row, gene of interest}
}
\description{
Fonction to barplot all the shortest paths between 1 gene of interest to all
metabolites in the data value in entry
}
\examples{
barplotFunction(hsa01100, data, gene)
}
\keyword{barplot,}
\keyword{igraphe,}
\keyword{shortestDistance}

