% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graph.R
\name{barplotFunction}
\alias{barplotFunction}
\title{function that output a barplot graph related to one specific gene with all
the shortest distances from that gene to all metabolites}
\usage{
barplotFunction(pathwayId, associatedGeneMetaDF, completeMetaboliteDF, gene)
}
\arguments{
\item{pathwayId,}{data(gene, metabolites), gene}
}
\description{
for param data:
     gene = KEGGid of gene hsa:...
     metabolites : KEGGid of metabolites C....
for param pathwayId : KEGG id of pathways without ':' ex: hsa01100
for param gene : is a gene in data ex: hsa:8801
}
\examples{
barplotFunction(hsa01100, associatedGeneMetaDF,
 completeMetaboliteDF, gene)
}
\keyword{Graph,}
\keyword{barplot,}
\keyword{shortestDistance}

