% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphe.R
\name{heatmapFunction}
\alias{heatmapFunction}
\title{HeatMap function, is using the result data.frame from getAllShortestDistances
to create it. It uses the associations data form the users.}
\usage{
heatmapFunction(pathwayId, data)
}
\arguments{
\item{keggpathwayId,}{data.frame() colunm c(gene, metabolites)
representing association between them for each row, gene of interest}
}
\description{
HeatMap function, is using the result data.frame from getAllShortestDistances
to create it. It uses the associations data form the users.
}
\examples{
heatmapFunction(hsa01100, data)
}
\keyword{heatmap,}
\keyword{igraphe,}
\keyword{shortestDistance}

